+++ CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased—presumed 0.5.1]
+
+### Added
+
+### Removed
+
+### Fixed
+
+- Make sure DuckDB extensions are loaded when opening the DuckDB connection
+
 ## [0.5.0] - 2025-11-03
 
 ### Added